feat: Add function to remove the nth node from the end of a linked list
This commit adds a new function, `removeNthFromEnd`, to the linked list module. The function removes the nth node from the end of a linked list. It uses two pointers, a fast pointer and a slow pointer, to traverse the list and find the node to be removed. The function also includes an alternative implementation, `removeNthFromEndNotOptimal`, which calculates the length of the list and then removes the node.

Closes #<issue_number>